Move main code out of __init__.py
